General idea of how to render
I want pixel-by-pixel control. To do this in SFML we could use a sf::Image, but to render it to the screen it would need to be rendered as a sf::Texture which involves a swap between RAM and VRAM. Using a sf::Texture directly saves this swap so is more performant. I will update the pixels in the pixel array, update the texture with this pixel array, then create a sprite from this texture and render it to the screen.