Clunky automatic zooming
Implemented automatic zoom that picks a quadrant of the screen to zoom into depending on mouse position. Not super smooth but works. Could make it smoother by reducing zoomPixels and frameLimit, and having some functionality to speed up zoom at the start when it's a bit slow? Click/hold multiplies zoomPixels by n?